Request fixes with user, middelwares and details
