+++ app/controllers/user.js
@@ -1,6 +1,7 @@
 const User = require('../models').user,
   saltRounds = 10,
   validations = require('./validations'),
+  tokens = require('../services/tokenGenerator'),
   bcrypt = require('bcryptjs'),
   logger = require('../logger'),
   errors = require('../errors');
@@ -26,6 +27,39 @@ const newUser = body => {
   }
 };
 
+exports.signIn = (req, res, next) => {
+  const singInUser = req.body
+    ? {
+        password: req.body.password,
+        email: req.body.email
+      }
+    : {};
+
+  const errMsg = validations.validateUser(singInUser);
+  if (errMsg.length > 0) {
+    next(errors.invalidUser(errMsg));
+  } else {
+    return User.searchUserWhere(['email', 'password'], { email: singInUser.email })
+      .then(userFound => {
+        if (userFound) {
+          return bcrypt.compare(singInUser.password, userFound.password).then(passwordValid => {
+            if (passwordValid) {
+              logger.info(`User ${userFound.email} singed in`);
+              const authentication = tokens.encode({ email: userFound.email });
+              res.set(tokens.header, authentication);
+              res.status(200).end();
+            } else {
+              next(errors.invalidUser(`Email ${singInUser.email} or password are incorrect.`));
+            }
+          });
+        } else {
+          next(errors.invalidUser(`Cannot find user ${singInUser.email} or is invalid.`));
+        }
+      })
+      .catch(next);
+  }
+};
+
 exports.createUser = (req, res, next) =>
   newUser(req.body)
     .then(user =>
@@ -38,18 +72,10 @@ exports.createUser = (req, res, next) =>
 
 exports.getListUsers = (req, res, next) => {
   const limit = req.query.limit || 10;
-  const offset = req.query.page || 0;
-  return User.findAndCountAll({
-    attributes: ['firstName', 'lastName', 'email'],
-    limit,
-    offset,
-    order: [['email', 'ASC']]
-  })
-    .catch(err => {
-      throw errors.databaseError(err.message);
-    })
-    .then(getUsers => {
-      res.send(getUsers);
+  const offset = req.query.page * limit || 0;
+  return User.paginationUsers(limit, offset)
+    .then(listUsers => {
+      res.send(listUsers);
     })
     .catch(next);
 };
