+++ app/routes.js
@@ -1,9 +1,8 @@
 const users = require('./controllers/user'),
-  validToken = require('./middlewares/validationTokens'),
-  userSingIn = require('./controllers/singUser');
+  validToken = require('./middlewares/validationTokens');
 
 exports.init = app => {
   app.post('/users', users.createUser);
-  app.post('/users/sessions', userSingIn.signIn);
+  app.post('/users/sessions', users.signIn);
   app.get('/users', [validToken.validateToken], users.getListUsers);
 };
